+++ src/Gate.IO.Api/Unified/Responses/GateUnifiedLoanRecord.cs
@@ -51,5 +51,6 @@ public record GateUnifiedLoanRecord
     /// Create time
     /// </summary>
     [JsonProperty("create_time")]
+    [JsonConverter(typeof(DateTimeConverter))]
     public DateTime CreateTime { get; set; }
 }